requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.19
botocore==1.24.19
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==0.10.0
s3transfer==0.5.2
sqlparse==0.4.2
whitenoise==6.0.0
